[CI] Removes type2 in process_registration and fix Windows GPU Reduced Ops CI Pipeline (#16530)
### Description
Windows GPU Reduced Ops CI Pipeline is broken due to the introduction of
a second template type in registered kernels. The python code checking
the registration is broken due to that. This PR addresses this issue on
the python side by keeping only one type equal to the concatenation of
the two types.